requirements.txt
langchain-community>=0.0.22
gradio>=4.26.0
faiss-cpu>=1.7.4
tiktoken>=0.6.0
langchain-ollama
#ollama>=0.1.6
python-dotenv>=1.0.0
pytest>=7.4.0
sentence-transformers>=2.2.2
tqdm>=4.66.1
